requirements.txt
pytest
sqlalchemy
python-dotenv
fastapi
uvicorn
httpx
prometheus-client
websockets
python-multipart
openai>=1.0.0
requests
fastapi-jwt-auth
